Replace BigDecimal/BigInteger by more commonly used classes (#160)
* replace BigInteger by int/Integer

* replace BigDecimal by double

* added CHANGELOG message

* removing redundant entries from the XSD files

---------

+++ api/src/main/java/info/rexs/schema/jaxb/Attribute.java
@@ -8,8 +8,6 @@
 
 package info.rexs.schema.jaxb;
 
-import java.math.BigDecimal;
-import java.math.BigInteger;
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlAttribute;
@@ -57,13 +55,13 @@ public class Attribute {
 
     protected EnumValues enumValues;
     @XmlAttribute(name = "numericId", required = true)
-    protected BigInteger numericId;
+    protected int numericId;
     @XmlAttribute(name = "attributeId", required = true)
     protected String attributeId;
     @XmlAttribute(name = "valueType", required = true)
-    protected BigInteger valueType;
+    protected int valueType;
     @XmlAttribute(name = "unit", required = true)
-    protected BigInteger unit;
+    protected int unit;
     @XmlAttribute(name = "symbol")
     protected String symbol;
     @XmlAttribute(name = "nameEn", required = true)
@@ -71,15 +69,15 @@ public class Attribute {
     @XmlAttribute(name = "nameDe", required = true)
     protected String nameDe;
     @XmlAttribute(name = "rangeMin")
-    protected BigDecimal rangeMin;
+    protected double rangeMin;
     @XmlAttribute(name = "rangeMinIntervalOpen")
     protected Boolean rangeMinIntervalOpen;
     @XmlAttribute(name = "rangeMax")
-    protected BigDecimal rangeMax;
+    protected double rangeMax;
     @XmlAttribute(name = "rangeMaxIntervalOpen")
     protected Boolean rangeMaxIntervalOpen;
     @XmlAttribute(name = "databaseId")
-    protected BigInteger databaseId;
+    protected int databaseId;
 
     /**
      * Ruft den Wert der enumValues-Eigenschaft ab.
@@ -110,10 +108,10 @@ public class Attribute {
      *
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public BigInteger getNumericId() {
+    public int getNumericId() {
         return numericId;
     }
 
@@ -122,10 +120,10 @@ public class Attribute {
      *
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public void setNumericId(BigInteger value) {
+    public void setNumericId(int value) {
         this.numericId = value;
     }
 
@@ -158,10 +156,10 @@ public class Attribute {
      *
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public BigInteger getValueType() {
+    public int getValueType() {
         return valueType;
     }
 
@@ -170,10 +168,10 @@ public class Attribute {
      *
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public void setValueType(BigInteger value) {
+    public void setValueType(int value) {
         this.valueType = value;
     }
 
@@ -182,10 +180,10 @@ public class Attribute {
      *
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public BigInteger getUnit() {
+    public int getUnit() {
         return unit;
     }
 
@@ -194,10 +192,10 @@ public class Attribute {
      *
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public void setUnit(BigInteger value) {
+    public void setUnit(int value) {
         this.unit = value;
     }
 
@@ -278,10 +276,10 @@ public class Attribute {
      *
      * @return
      *     possible object is
-     *     {@link BigDecimal }
+     *     {@code double}
      *
      */
-    public BigDecimal getRangeMin() {
+    public double getRangeMin() {
         return rangeMin;
     }
 
@@ -290,10 +288,10 @@ public class Attribute {
      *
      * @param value
      *     allowed object is
-     *     {@link BigDecimal }
+     *     {@code double}
      *
      */
-    public void setRangeMin(BigDecimal value) {
+    public void setRangeMin(double value) {
         this.rangeMin = value;
     }
 
@@ -326,10 +324,10 @@ public class Attribute {
      *
      * @return
      *     possible object is
-     *     {@link BigDecimal }
+     *     {@code double}
      *
      */
-    public BigDecimal getRangeMax() {
+    public double getRangeMax() {
         return rangeMax;
     }
 
@@ -338,10 +336,10 @@ public class Attribute {
      *
      * @param value
      *     allowed object is
-     *     {@link BigDecimal }
+     *     {@code double}
      *
      */
-    public void setRangeMax(BigDecimal value) {
+    public void setRangeMax(double value) {
         this.rangeMax = value;
     }
 
@@ -374,10 +372,10 @@ public class Attribute {
      *
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public BigInteger getDatabaseId() {
+    public int getDatabaseId() {
         return databaseId;
     }
 
@@ -386,10 +384,10 @@ public class Attribute {
      *
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@code int}
      *
      */
-    public void setDatabaseId(BigInteger value) {
+    public void setDatabaseId(int value) {
         this.databaseId = value;
     }
